+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.2] 2021-11-16
+
+- Add `url`, `regex` and `ip` validators.
+
+## [3.0.1] 2021-11-17
+
+- Remove the `ctx` parameter of `CustomValidator::check`. [#710](https://github.com/async-graphql/async-graphql/issues/710)
+
 ## [3.0.0-alpha.2] 2021-11-16
 
 - Change the signature of the `connection::query` function to allow the callback to use any type that implements `Into<Error>`.